requirements.txt
# Core
streamlit
python-dotenv
openai-agents
google-generativeai

# OpenAI LLM support
openai

# Google Calendar API
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

pydantic
nest-asyncio

openai-whisper
streamlit-webrtc
soundfile
av
